+++ server/src/main/java/org/realityforge/replicant/server/ee/ReplicationRequestUtil.java
@@ -29,13 +29,13 @@ public final class ReplicationRequestUtil
   {
   }
 
-  static <T> T runRequest( @Nonnull final TransactionSynchronizationRegistry registry,
-                           @Nonnull final EntityManager entityManager,
-                           @Nonnull final EntityMessageEndpoint endpoint,
-                           @Nonnull final String invocationKey,
-                           @Nullable final String sessionID,
-                           @Nullable final String requestID,
-                           @Nonnull final Callable<T> action )
+  public static <T> T runRequest( @Nonnull final TransactionSynchronizationRegistry registry,
+                                  @Nonnull final EntityManager entityManager,
+                                  @Nonnull final EntityMessageEndpoint endpoint,
+                                  @Nonnull final String invocationKey,
+                                  @Nullable final String sessionID,
+                                  @Nullable final String requestID,
+                                  @Nonnull final Callable<T> action )
     throws Exception
   {
     startReplication( registry, invocationKey, sessionID, requestID );